+++ src/ir/global_initialization.mli
@@ -1,5 +1,7 @@
 type phase = Compile_initializer | Load_initializer
 type region
+type static_region
+type storage_region
 type t
 
 type region_description = {
@@ -8,7 +10,15 @@ type region_description = {
   last : Instruction_sequence.Instruction_id.t;
 }
 
+type static_region_description = {
+  static_root : Sema.Function_call_expression_result.initializer_result;
+  static_slot : Integer_globals.static_slot;
+  first : Instruction_sequence.Instruction_id.t;
+  last : Instruction_sequence.Instruction_id.t;
+}
+
 val create :
+  ?static_descriptions:static_region_description list ->
   span:Common.Span.t ->
   globals:Integer_globals.t ->
   entry:X87_stack.t ->
@@ -23,6 +33,27 @@ val create :
 
 val matches : t -> globals:Integer_globals.t -> entry:X87_stack.t -> bool
 val regions : t -> region list
+val static_regions : t -> static_region list
+
+val static_root :
+  static_region -> Sema.Function_call_expression_result.initializer_result
+
+val static_slot : static_region -> Integer_globals.static_slot
+val describe_static : static_region -> static_region_description
+val static_phase : static_region -> phase
+val storage_regions : t -> storage_region list
+
+val find_storage :
+  t -> Instruction_sequence.Instruction_id.t -> storage_region option
+
+val storage_symbol : storage_region -> Sema.Symbol.t
+val storage_phase : storage_region -> phase
+
+val storage_frame :
+  storage_region -> Sema.Function_frame_layout.function_layout option
+
+val storage_first : storage_region -> Instruction_sequence.Instruction_id.t
+val storage_last : storage_region -> Instruction_sequence.Instruction_id.t
 val prepared_steps : t -> int
 val find : t -> Instruction_sequence.Instruction_id.t -> region option
 val root : region -> Sema.Function_call_expression_result.top_level_root_result